metrics: turn error codes into constants
